transfer to integer solution

+++ src/tracks/singlePlayer/ECAssignment3/controllers/betterRHEA/HyperParamSet.java
@@ -15,12 +15,12 @@ public class HyperParamSet {
     public final int INDIVIDUAL_DEPTH;    // 0 - 1.0 , gamma / 1000, 0 - 1001
     public final int POPULATION_SIZE;  // 1 - 20
     public final int TOURNAMENT_SIZE;   // 2 - 20
-    public final double CROSSOVER_RATE,MUTATION_RATE,WIN_BONUS,LOSE_PENALTY;
+    public final int CROSSOVER_RATE,MUTATION_RATE,WIN_BONUS,LOSE_PENALTY;
 
-    public HyperParamSet(double INDIVIDUAL_DEPTH, double POPULATION_SIZE, double TOURNAMENT_SIZE, double CROSSOVER_RATE,double MUTATION_RATE, double WIN_BONUS,double LOSE_PENALTY) {
-        this.INDIVIDUAL_DEPTH = (int)INDIVIDUAL_DEPTH;
-        this.POPULATION_SIZE = (int)POPULATION_SIZE;
-        this.TOURNAMENT_SIZE = (int)TOURNAMENT_SIZE;
+    public HyperParamSet(int INDIVIDUAL_DEPTH, int POPULATION_SIZE, int TOURNAMENT_SIZE, int CROSSOVER_RATE,int MUTATION_RATE, int WIN_BONUS,int LOSE_PENALTY) {
+        this.INDIVIDUAL_DEPTH = INDIVIDUAL_DEPTH;
+        this.POPULATION_SIZE = POPULATION_SIZE;
+        this.TOURNAMENT_SIZE = TOURNAMENT_SIZE;
         this.CROSSOVER_RATE = CROSSOVER_RATE;
         this.MUTATION_RATE =MUTATION_RATE;
         this.WIN_BONUS=WIN_BONUS;
